pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "writerbox"
version = "0.1.0-alpha"
description = "A terminal-based writing collection manager"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Brennan Brown", email = "brennan@ BrennanBrown.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Editors",
    "Topic :: Utilities",
]
requires-python = ">=3.9"
dependencies = [
    "textual>=0.41.0",
    "click>=8.0.0",
    "python-frontmatter>=1.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "markdown>=3.4.0",
    "pygments>=2.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
writerbox = "writerbox.cli:main"

[project.urls]
Homepage = "https://github.com/brennanbrown/writerbox"
Repository = "https://github.com/brennanbrown/writerbox"
Documentation = "https://github.com/brennanbrown/writerbox#readme"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
